GeometryTypes - Operations

GeometryTypes - Operations

Slicing

GeometryTypes.slice — Function.

Slice an AbstractMesh at the specified Z axis value. Returns a Vector of LineSegments generated from the faces at the specified heights. Note: This will not slice in-plane faces.

source

Set Operations

Base.union — Function.

Perform a union between two HyperRectangles.

source

Base.diff — Function.

Perform a difference between two HyperRectangles.

source

Base.intersect — Function.

Perform a intersection between two HyperRectangles.

source

Decompositions

GeometryTypes.decompose — Function.

Allow to call decompose with unspecified vector type and infer types from primitive.

source

decompose{N, FT1, FT2, O1, O2}(::Type{Face{3, FT1, O1}},
                               f::Face{N, FT2, O2})

Triangulate an N-Face into a tuple of triangular faces.

source

decompose{N, FT1, FT2, O1, O2}(::Type{Face{2, FT1, O1}},
                               f::Face{N, FT2, O2})

Extract all line segments in a Face.

source

Decompose an N-Simplex into a tuple of Simplex{3}

source

Decompose an N-Simplex into tuple of Simplex{2}

source

Decompose an N-Simplex into a tuple of Simplex{1}

source

Get decompose a HyperRectangle into points.

source

Get decompose a HyperRectangle into Texture Coordinates.

source

Get decompose a HyperRectangle into faces.

source